requirements.txt
streamlit
google-generativeai
ephem
geopy
pandas